% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterCWS}
\alias{filterCWS}
\title{filter cws data}
\usage{
filterCWS(data, m1_cutOff = 1, m2_low = 0.1, m2_high = 0.95,
  m3_cutOff = 0.2, m4_cutOff = 0.9, o1_fun = interpol, o2_cutOff = 0.8,
  o3_cutOff = 0.8, includeOptional = T, ...)
}
\arguments{
\item{data}{input data in a format like data(netatmoBer)}

\item{m1_cutOff}{see cutoff in ?m1}

\item{m2_low}{see low in ?m2}

\item{m2_high}{see high in ?m2}

\item{m3_cutOff}{see cutoff in ?m3}

\item{m4_cutOff}{see cutoff in ?m4}

\item{o1_fun}{see fun in ?o1}

\item{o2_cutOff}{see cutoff in ?o2}

\item{o3_cutOff}{see cutoff in ?o3}

\item{includeOptional}{if the steps o1 to o3 should be performed}

\item{...}{additional parameters used in o1 for details see ?o1}
}
\value{
data.table
}
\description{
performs all filter steps.
}
\examples{
data(netatmoBer)
y <- filterCWS(netatmoBer)
}
