% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getZ}
\alias{getZ}
\title{calculate robust z-score}
\usage{
getZ(x)
}
\arguments{
\item{x}{vector to caluclate the robust score from}
}
\value{
vector with z-Score for elements in x or NaN
}
\description{
calculate robust z-score
}
