% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{interpol}
\alias{interpol}
\title{interpol}
\usage{
interpol(x, maxLength = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{maxLength}{How long a gap can be before it remains}
}
\value{
vector
}
\description{
This function takes a numerical vector x and fills NaN's with linear
interpolated values. But only if the NaN gap is smaller or equal maxLength
}
\examples{
x <- x(1, NaN, 3, NaN NaN, 6, NaN, 8)
interpol(x)
interpol(x, 2)
}
