% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{m1}
\alias{m1}
\title{Main QC step m1}
\usage{
m1(data, cutOff = 1)
}
\arguments{
\item{data}{data set formated as the sample data (netatmoBer)}

\item{cutOff}{how much stations are allowed to have the same coordinates,
default is 1. This means that if two p_ids share the same lat lon values,
data at these stations are set to FALSE.}
}
\value{
data.table
}
\description{
Flag values with FALSE that are missing or belong to a p_id having lon lat
values which occur more than the 'cutOff' value.
}
\examples{
data(netatmoBer)
x <- m1(netatmoBer)
}
