% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{m4}
\alias{m4}
\title{Main QC step m4}
\usage{
m4(data, cutOff = 0.9)
}
\arguments{
\item{data}{data.table as returned by m3}

\item{cutOff}{value of correlation coefficient below which data are flagged
with FALSE, 0 < cutOff < 1. Default is 0.9.}
}
\value{
data.table
}
\description{
Flag values with FALSE if they belong to a month in which the correlation
with the median of all stations is lower than cutOff.
}
\examples{
y <- m3(x)
z <- m4(y)
}
