% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{o2}
\alias{o2}
\title{filter step o2}
\usage{
o2(data, cutOff = 0.8)
}
\arguments{
\item{data}{data.table as returned from o1}

\item{cutOff}{percentage of values that must be present before all values of
a day are flagged with false}
}
\value{
data.table
}
\description{
For consitency, filter if lesser than 80 percent of values for a day are
present. Steps can be skipped by renaming columns in the input data.
}
\examples{
o_2 <- o2(o_1)
}
