% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{Qnr}
\alias{Qnr}
\title{Version of Qn respecting NaN values}
\usage{
Qnr(x)
}
\arguments{
\item{x}{}
}
\value{
NaN for no valid values otherwise Qn without NaN
}
\description{
Version of Qn respecting NaN values
}
