% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getZ}
\alias{getZ}
\title{Calculate robust z-score}
\usage{
getZ(x)
}
\arguments{
\item{x}{vector to calculate the robust z-score from}
}
\value{
vector with z-score for elements in x or NaN
}
\description{
Calculate robust z-score
}
