% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{m1}
\alias{m1}
\title{perform filter step m1}
\usage{
m1(data, cutOff = 1)
}
\arguments{
\item{data}{data set formated as netatmoBer}

\item{cutOff}{how much stations are allowed to have the same coordinates,
default is 1 meaning no p_id's with the same lat lon values are allowed}
}
\value{
data.table
}
\description{
flag values with false that are missing or belong to a p_id having lon lat values
which appear more than cutOff
}
\examples{
data(netatmoBer)
x <- m1(netatmoBer)
}
