% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{m2}
\alias{m2}
\title{perform filter step m2}
\usage{
m2(data, low = 0.01, high = 0.95, debug = F)
}
\arguments{
\item{data}{data.table object obtained from m1}

\item{low}{value > 0}

\item{high}{value < 1}

\item{debug}{set to true too keep intermediate results}
}
\value{
data.table
}
\description{
flages all values which robust z-score is not within of the critical values
obtained from low and high
Steps can be skipped by renaming columns in the input data.
}
\examples{
data(netatmoBer)
x <- m1(netatmoBer)
y <- m2(x)

}
