% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{m2}
\alias{m2}
\title{Main QC step m2}
\usage{
m2(data, low = 0.01, high = 0.95, heightCorrection = T, debug = F)
}
\arguments{
\item{data}{data.table object obtained from m1}

\item{low}{0 < low < high < 1}

\item{high}{0 < low < high < 1}

\item{heightCorrection}{if set to true (default) and the column "z" exists in
the input data, the temperatures used in calculating the the z-score are
corrected. The applied formula is t_cor = t + ((0.0065 * (z - mz)) where
mz is the spatial mean at each time step}

\item{debug}{set to true to keep intermediate results}
}
\value{
data.table
}
\description{
Flags all values as FALSE if robust z-score is not within the critical values
obtained from low and high. This approach is based on the distribution
function of all values. !Careful: if number of stations is < 200, it would be
better to use the Student t distribution (to be implemented)!
}
\examples{
data(netatmoBer)
x <- m1(netatmoBer)
y <- m2(x)

}
