% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{m4}
\alias{m4}
\title{m4}
\usage{
m4(data, cutOff = 0.9)
}
\arguments{
\item{data}{data.table as returned by m3}

\item{cutOff}{}
}
\value{
data.table
}
\description{
flag values with false if they belong to a month were the correlation with
median of all station is lower than cutOff. Steps can be skipped by renaming
columns in the input data.
}
\examples{
y <- m3(x)
z <- m4(y)
}
