% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{o3}
\alias{o3}
\title{step o3}
\usage{
o3(data, cutOff = 0.8)
}
\arguments{
\item{data}{data.table as returned from o2}

\item{cutOff}{percentage of values that ust be present before the month is flagged false}
}
\value{
data.table
}
\description{
For consitency filter if lesse than 80 percent of values for a day are
present. Steps can be skipped by renaming columns in the input data.
}
\examples{
o_3 <- o3(o_2)
}
